feat: Add useFrame hook
