+++ packages/insight-viewer/src/hooks/useMultiframeImages/index.ts
@@ -2,9 +2,9 @@
  * @fileoverview Loads images(Dicom/Web) and
  * return the current image and loading state, current frame and frame updater.
  */
-import { SetStateAction } from 'react'
+
 import { useLoadImages } from './useLoadImages'
-import useFrame, { SetFrame } from './useFrame'
+
 import { HTTP, LoaderType } from '../../types'
 import { LOADER_TYPE, CONFIG } from '../../const'
 import { ImagesLoadState } from './types'
@@ -16,10 +16,13 @@ type Prop = {
 } & Partial<HTTP>
 
 interface UseMultiframeImages {
-  ({ imageIds, initialFrame, onError, requestInterceptor, type }: Prop): {
-    frame: number // current frame index
-    setFrame: SetFrame // set current frame index
-  } & ImagesLoadState
+  ({
+    imageIds,
+    initialFrame,
+    onError,
+    requestInterceptor,
+    type,
+  }: Prop): ImagesLoadState
 }
 
 /**
@@ -35,7 +38,6 @@ interface UseMultiframeImages {
 export const useMultiframeImages: UseMultiframeImages = ({
   imageIds,
   type = LOADER_TYPE.Dicom,
-  initialFrame = 0,
   requestInterceptor = CONFIG.requestInterceptor,
   onError = CONFIG.onError,
 }) => {
@@ -46,24 +48,7 @@ export const useMultiframeImages: UseMultiframeImages = ({
     type,
   })
 
-  const { frame, setFrame } = useFrame(initialFrame ?? 0)
-
-  function handleFrame(arg: SetStateAction<number>) {
-    let frameIndex = 0
-
-    if (typeof arg === 'number') {
-      frameIndex = arg
-    } else {
-      frameIndex = arg(frame)
-    }
-
-    if (frameIndex < 0 || frameIndex > imageIds.length - 1) return
-    setFrame(arg)
-  }
-
   return {
-    frame,
-    setFrame: handleFrame,
     loadingStates,
     images: loadedImages,
   }
