+++ packages/insight-viewer/src/index.ts
@@ -3,8 +3,9 @@ export { useMultiframeImages } from './hooks/useMultiframeImages'
 export { useViewport } from './hooks/useViewport'
 export { useInteraction } from './hooks/useInteraction'
 export { useImage } from './hooks/useImage'
-export type { Interaction, SetInteraction } from './hooks/useInteraction/types'
+export { useFrame } from './hooks/useFrame'
 export { isValidViewport } from './utils/common'
+export type { Interaction, SetInteraction } from './hooks/useInteraction/types'
 export type { Viewport, ViewerError } from './types'
 export type {
   DragEvent,
